requirements.txt
annotated-types==0.7.0
anyio==4.5.2
cachecontrol==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
exceptiongroup==1.2.2
fastapi==0.115.4
firebase-admin==6.5.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
PyJWT==2.9.0
pyparsing==3.1.4
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.41.2
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0